+++ src/framework/Vector2.cpp
@@ -9,7 +9,7 @@ geo::framework::Vector2::Vector2()
     : x(0)
     , y(0) { }
 
-geo::framework::Vector2::Vector2(float x, float y)
+geo::framework::Vector2::Vector2(int32_t x, int32_t y)
     : x(x)
     , y(y) { }
 
@@ -68,7 +68,7 @@ geo::framework::Vector2 &geo::framework::Vector2::operator/=(const geo::framewor
 void geo::framework::Vector2::add(const geo::framework::Vector2 &other) {
     add(other.x, other.y);
 }
-void geo::framework::Vector2::add(float x, float y) {
+void geo::framework::Vector2::add(int32_t x, int32_t y) {
     this->x += x;
     this->y += y;
 }
@@ -76,7 +76,7 @@ void geo::framework::Vector2::add(float x, float y) {
 void geo::framework::Vector2::sub(const geo::framework::Vector2 &other) {
     sub(other.x, other.y);
 }
-void geo::framework::Vector2::sub(float x, float y) {
+void geo::framework::Vector2::sub(int32_t x, int32_t y) {
     this->x -= x;
     this->y -= y;
 }
@@ -84,7 +84,7 @@ void geo::framework::Vector2::sub(float x, float y) {
 void geo::framework::Vector2::mul(const geo::framework::Vector2 &other) {
     mul(other.x, other.y);
 }
-void geo::framework::Vector2::mul(float x, float y) {
+void geo::framework::Vector2::mul(int32_t x, int32_t y) {
     this->x *= x;
     this->y *= y;
 }
@@ -92,28 +92,28 @@ void geo::framework::Vector2::mul(float x, float y) {
 void geo::framework::Vector2::div(const geo::framework::Vector2 &other) {
     div(other.x, other.y);
 }
-void geo::framework::Vector2::div(float x, float y) {
+void geo::framework::Vector2::div(int32_t x, int32_t y) {
     this->x /= x;
     this->y /= y;
 }
 
 
-float geo::framework::Vector2::getX() const {
+uint32_t geo::framework::Vector2::getX() const {
     return x;
 }
-void geo::framework::Vector2::setX(float x) {
+void geo::framework::Vector2::setX(int32_t x) {
     this->x = x;
 }
 
-float geo::framework::Vector2::getY() const {
+uint32_t geo::framework::Vector2::getY() const {
     return y;
 }
-void geo::framework::Vector2::setY(float y) {
+void geo::framework::Vector2::setY(int32_t y) {
     this->y = y;
 }
 
 float geo::framework::Vector2::getLength() const {
-    return sqrtf(x * x + y * y);
+    return sqrtf(static_cast<float>(x * x + y * y));
 }
